fix: move logic from route to page folder (#125)
* removed logic from route

* added new pages

* fix build errors
